requirements.txt
django
channels
websockets
python-binance
pytest